refactor: move footer to layout folder
